+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased] 
+------------------------
+##### Changed
+- Vagrant setup script
+
 [1.11.1] - 2015-06-29
 ------------------------
 ##### Fixed